update: ubuntuやdebianのときにaliasとしてuag,aucを設定するように変更

+++ .zshrc
@@ -41,12 +41,14 @@ setopt extended_history
 
 # alias
 if type -p "eza" > /dev/null 2>&1; then
-    alias ll='eza -aal -F=always --icons=always --group-directories-first'
+  alias ll='eza -aal -F=always --icons=always --group-directories-first'
 else
-    alias ll='ls -alF'
+  alias ll='ls -alF'
 fi
-alias uag='sudo apt update && sudo apt upgrade -y'
-alias auc='sudo apt autoremove && sudo apt clean -y'
+if [[ -e /etc/debian_version ]]; then
+  alias uag='sudo apt update && sudo apt upgrade -y'
+  alias auc='sudo apt autoremove && sudo apt clean -y'
+fi
 alias nv='nvim'
 alias nz='nvim ~/.zshrc'
 alias nzl='nvim ~/.zshrc.local'
